Refactor Viewsets (#345)
* Refactor Viewsets
