+++ data_portal/viewsets/run.py
@@ -1,0 +1,33 @@
+# -*- coding: utf-8 -*-
+"""viewsets module
+
+NOTE:
+     This is DRF based Portal API impls.
+"""
+import logging
+
+from rest_framework import filters
+from rest_framework.response import Response
+from rest_framework.viewsets import ReadOnlyModelViewSet
+
+from data_portal.models.limsrow import LIMSRow
+from data_portal.pagination import StandardResultsSetPagination
+from data_portal.serializers import RunIdSerializer
+
+logger = logging.getLogger()
+
+
+class RunViewSet(ReadOnlyModelViewSet):
+    queryset = LIMSRow.objects.values_list('illumina_id', named=True).filter(illumina_id__isnull=False).distinct()
+    serializer_class = RunIdSerializer
+    pagination_class = StandardResultsSetPagination
+    filter_backends = [filters.OrderingFilter, filters.SearchFilter]
+    ordering_fields = ['illumina_id']
+    ordering = ['-illumina_id']
+    search_fields = ordering_fields
+
+    def retrieve(self, request, pk=None, **kwargs):
+        data = {
+            'id': pk,
+        }
+        return Response(data)
